Split init module into chain and wallet-related parts.
