+++ zcash_client_sqlite/src/wallet/init.rs
@@ -10,33 +10,7 @@ use zcash_primitives::{
 
 use zcash_client_backend::{data_api::error::Error, encoding::encode_extended_full_viewing_key};
 
-use crate::{address_from_extfvk, error::SqliteClientError, CacheConnection, DataConnection};
-
-/// Sets up the internal structure of the cache database.
-///
-/// # Examples
-///
-/// ```
-/// use tempfile::NamedTempFile;
-/// use zcash_client_sqlite::{
-///     CacheConnection,
-///     init::init_cache_database,
-/// };
-///
-/// let cache_file = NamedTempFile::new().unwrap();
-/// let db = CacheConnection::for_path(cache_file.path()).unwrap();
-/// init_cache_database(&db).unwrap();
-/// ```
-pub fn init_cache_database(db_cache: &CacheConnection) -> Result<(), rusqlite::Error> {
-    db_cache.0.execute(
-        "CREATE TABLE IF NOT EXISTS compactblocks (
-            height INTEGER PRIMARY KEY,
-            data BLOB NOT NULL
-        )",
-        NO_PARAMS,
-    )?;
-    Ok(())
-}
+use crate::{address_from_extfvk, error::SqliteClientError, DataConnection};
 
 /// Sets up the internal structure of the data database.
 ///
